requirements.txt
certifi==2018.8.24
chardet==3.0.4
gitdb2==2.0.4
GitPython==2.1.11
idna==2.7
requests==2.19.1
smmap2==2.0.4
urllib3==1.23
